Fenced Code Block Slide Template
Figuring out how remark accepts fenced coding blocks, and what I can do to manipulate them.

If adding a background image with the same color code as the fenced blocks, I can play with the position in absolute.

Since this might eliminate the header of the slide, it might be necessary to have a title slide right before it.