Clean Up
Realised that I did some stuff that was redundant, so I went back and cleaned up the things that didn't need to be there.

Some of these things being: Footer.png where it was not needed, completely took out the horizontalBar.png because it could be replaced with border-left or border-right.

This allowed for more dynamic pages, where images can just be dropped in and text can just be written in and the class:center,middle will do the rest.